Refactor: Centralize Quarto chunk generation in R6 class
This commit refactors the report generation process by moving logic from external scripts into the `NeuropsychResultsR6` class.

A new `emit_quarto_text_chunk` method has been added to the R6 class. This method is responsible for generating the complete Quarto R code chunk required to process data for a specific domain. This centralizes the logic and ensures consistency.

As a result of this change:
- The `create_text_placeholder` method is simplified, now only creating an empty file if one doesn't exist.
- The now-obsolete workflow scripts (`neuro2_R6_update_workflow.R` and `neuro2_duckdb_workflow_DEMO.R`) have been removed.

Signed-off-by: brainworkup <j.trampush@gmail.com>